Added notes to test method
Removed test case for Create_ShouldFail_WhenBookingConflictsWithExistingBooking

This is because this test conflicts with the M3 booking logic. We might need to implement this logic in M4.
M3 Fact: A vehicle in a pod is available for booking.
M3 Fact: If a vehicle has-a booking, then it does not have-a pod.